+++ src/parser/inline/media.test.ts
@@ -11,10 +11,10 @@ describe('Unit: parser/inline/media', () => {
       assert.deepStrictEqual(inspect(parser('![]{vbscript:alert}')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{data-type="image" data:text/html;base64,PHNjcmlwdD5hbGVydCgnWFNTJyk8L3NjcmlwdD4K}')), undefined);
       assert.deepStrictEqual(inspect(parser('![]{any:alert}')), undefined);
-      assert.deepStrictEqual(inspect(parser('![]{"}')), [['<a href="&quot;" rel="noopener"><img class="media" data-src="&quot;" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{\\}')), [['<a href="\\" rel="noopener"><img class="media" data-src="\\" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('!["]{/}')), [['<a href="/" rel="noopener"><img class="media" data-src="/" alt="&quot;"></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![\\"]{/}')), [['<a href="/" rel="noopener"><img class="media" data-src="/" alt="&quot;"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{"}')), [['<a href="&quot;" rel="noopener" target="_blank"><img class="media" data-src="&quot;" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{\\}')), [['<a href="\\" rel="noopener" target="_blank"><img class="media" data-src="\\" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!["]{/}')), [['<a href="/" rel="noopener" target="_blank"><img class="media" data-src="/" alt="&quot;"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![\\"]{/}')), [['<a href="/" rel="noopener" target="_blank"><img class="media" data-src="/" alt="&quot;"></a>'], '']);
     });
 
     it('invalid', () => {
@@ -49,26 +49,26 @@ describe('Unit: parser/inline/media', () => {
     });
 
     it('basic', () => {
-      assert.deepStrictEqual(inspect(parser('![]{b}')), [['<a href="b" rel="noopener"><img class="media" data-src="b" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{b }')), [['<a href="b" rel="noopener"><img class="media" data-src="b" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{ b }')), [['<a href="b" rel="noopener"><img class="media" data-src="b" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{\\}')), [['<a href="\\" rel="noopener"><img class="media" data-src="\\" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{\\ }')), [['<a href="\\" rel="noopener"><img class="media" data-src="\\" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{\\b}')), [['<a href="\\b" rel="noopener"><img class="media" data-src="\\b" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{a b}')), [['<a href="a" rel="noopener" class="invalid" data-invalid-syntax="link" data-invalid-type="parameter"><img class="media invalid" data-src="a" alt="" data-invalid-syntax="media" data-invalid-type="parameter"></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![a ]{b}')), [['<a href="b" rel="noopener"><img class="media" data-src="b" alt="a"></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![a b]{c}')), [['<a href="c" rel="noopener"><img class="media" data-src="c" alt="a b"></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('!{b}')), [['<a href="b" rel="noopener"><img class="media" data-src="b" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('!{ ][ }')), [['<a href="][" rel="noopener"><img class="media" data-src="][" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('!{ }{ }')), [['<a href="}{" rel="noopener"><img class="media" data-src="}{" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{b}')), [['<a href="b" rel="noopener" target="_blank"><img class="media" data-src="b" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{b }')), [['<a href="b" rel="noopener" target="_blank"><img class="media" data-src="b" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{ b }')), [['<a href="b" rel="noopener" target="_blank"><img class="media" data-src="b" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{\\}')), [['<a href="\\" rel="noopener" target="_blank"><img class="media" data-src="\\" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{\\ }')), [['<a href="\\" rel="noopener" target="_blank"><img class="media" data-src="\\" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{\\b}')), [['<a href="\\b" rel="noopener" target="_blank"><img class="media" data-src="\\b" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{a b}')), [['<a href="a" rel="noopener" class="invalid" data-invalid-syntax="link" data-invalid-type="parameter" target="_blank"><img class="media invalid" data-src="a" alt="" data-invalid-syntax="media" data-invalid-type="parameter"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![a ]{b}')), [['<a href="b" rel="noopener" target="_blank"><img class="media" data-src="b" alt="a"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![a b]{c}')), [['<a href="c" rel="noopener" target="_blank"><img class="media" data-src="c" alt="a b"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!{b}')), [['<a href="b" rel="noopener" target="_blank"><img class="media" data-src="b" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!{ ][ }')), [['<a href="][" rel="noopener" target="_blank"><img class="media" data-src="][" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!{ }{ }')), [['<a href="}{" rel="noopener" target="_blank"><img class="media" data-src="}{" alt=""></a>'], '']);
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('![]{{b}}')), [['<a href="{b}" rel="noopener"><img class="media" data-src="{b}" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{{a{b}{}{c}d}}')), [['<a href="{a{b}{}{c}d}" rel="noopener"><img class="media" data-src="{a{b}{}{c}d}" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{ }{ }')), [['<a href="}{" rel="noopener"><img class="media" data-src="}{" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![\\[]{/}')), [['<a href="/" rel="noopener"><img class="media" data-src="/" alt="["></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![<wbr>"]{"?"#"}')), [['<a href="&quot;?&quot;#&quot;" rel="noopener"><img class="media" data-src="&quot;?&quot;#&quot;" alt="<wbr>&quot;"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{{b}}')), [['<a href="{b}" rel="noopener" target="_blank"><img class="media" data-src="{b}" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{{a{b}{}{c}d}}')), [['<a href="{a{b}{}{c}d}" rel="noopener" target="_blank"><img class="media" data-src="{a{b}{}{c}d}" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{ }{ }')), [['<a href="}{" rel="noopener" target="_blank"><img class="media" data-src="}{" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![\\[]{/}')), [['<a href="/" rel="noopener" target="_blank"><img class="media" data-src="/" alt="["></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![<wbr>"]{"?"#"}')), [['<a href="&quot;?&quot;#&quot;" rel="noopener" target="_blank"><img class="media" data-src="&quot;?&quot;#&quot;" alt="<wbr>&quot;"></a>'], '']);
     });
 
     it('external', () => {
@@ -77,9 +77,9 @@ describe('Unit: parser/inline/media', () => {
     });
 
     it('attribute', () => {
-      assert.deepStrictEqual(inspect(parser('![]{/ nofollow}')), [['<a href="/" rel="noopener nofollow noreferrer"><img class="media" data-src="/" alt=""></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('![]{/ constructor}')), [['<a href="/" rel="noopener" class="invalid" data-invalid-syntax="link" data-invalid-type="parameter"><img class="media invalid" data-src="/" alt="" data-invalid-syntax="media" data-invalid-type="parameter"></a>'], '']);
-      assert.deepStrictEqual(inspect(parser('!{/ nofollow}')), [['<a href="/" rel="noopener nofollow noreferrer"><img class="media" data-src="/" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{/ nofollow}')), [['<a href="/" rel="noopener nofollow noreferrer" target="_blank"><img class="media" data-src="/" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{/ constructor}')), [['<a href="/" rel="noopener" class="invalid" data-invalid-syntax="link" data-invalid-type="parameter" target="_blank"><img class="media invalid" data-src="/" alt="" data-invalid-syntax="media" data-invalid-type="parameter"></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!{/ nofollow}')), [['<a href="/" rel="noopener nofollow noreferrer" target="_blank"><img class="media" data-src="/" alt=""></a>'], '']);
     });
 
   });